Read in the useAsTargetIdentifier from the workflow definition. This affects a number procs as we need to chase this value through from proc to proc.
Basically what should now happen is for a stored data element if that element is marked as the target identifier then it will update the target caption in ASRSysWorkflowInstances. Likewise for a record selector if its marked then it will update the caption.

Note, that at this stage the caption is initially populated with the instanciating record when the workflow is kicked off. This will eventually be 'unidentifed' if something else in the workflow is marked at the target. Similarily there is as yet no validation on stopping a user define multiple use as target identifier. These will be addressed in subsequent changesets.

There are a number of procs in this changeset that have not changed, but have changed location in the project file. This is because they were in da wrong place! There is no need to add these to the update script, as this is merely a Source control tidying up exercise.