Problems with the expiry date if the region was not UK English.
Cutting a long story short its to do with Lotus 123. No seriously it is. Way back in the 80s Lotus 123 had a bug whereby they forgot that 1900 was a leap year and so all of their date functions were based on this. When Microsoft brought out Excel they were desperate to grab the Lotus customers and so came up with a promise that their spreadsheets would work seamlessly - they therefore deliberately copied the bug. You can see this if you go into Excel and run some date functions on the 31/12/1899. It won't play ball. As far as Microsoft is concerned this date does not exist. 

So, what's this got to do with this bug I hear you ask. Well, MSoft in their wisdom copied this code into their vbScript, and subsequently into their VB code. They also decided that date variables can not be null, and so when I want to store a empty date I set it to - Wait for it - 31/12/1899. Code no like that. To compound problems American date format returns the time component (even though we're not using it) as 12:00:00AM, whereas uk returns 00:00:00, and so my empty checks were not kicking in, and then the check was happening for a date that does not exist. (Apparently 29th Feb 1904 is a problematic date too)

Ok, so that wasn't a short story. The bottom line is that I've replaced the date variable with a numeric one, which takes us nicely to 31/12/9999, but can cause funny results if you put a date in before 1901 (it'll detect the licence has expired, but the display in system mmanager might be wrong. Does it matter? If someone is going to put in a licence with an expiry of over a hundred years ago they deserve everything they get - which is nothing!

One final point, the licence code generator has changed so if you run this on an 8.1 database without regenerating teh code the expiry date will be calualted using the julian date format and hence will give you wacky expiry dates (the current func 8.1 db will expiry sometime in 7100)