Please forgive some of this code.
The Nine box grid uses the same engine as the cross tab reports, but with two key differences. Firstly the vertical axis is upside down for nine boxes, i.e. the higher values appear at the top of the grid.

The second one is how the step increments are calculated. In cross tabs its up to the user to enter an increment, but with nineboxes we always calculate. This may appear to be a trivial point, but, and its a big but, we need to ensure a nice tight fit of the range within the 3 rows or columns of the nine box, which is not always easy when we have a integer range, or something that doesn't snugly divide by 3.

Reverted part fo the getGroupNuber to before the initial flip of the data upsde down to ensure cross tabs are not affected.

We no longer flip the max and min of the vertcial axis as part of the retrivedefiniton, but instead flip the data arrays later. This may be a bit ugly (it IS ugly) but it means we don't have different processing in the loops that calulate which group a record is in.

This has only been written with up to 2 decimals places in mind. Because of discrepancies between sql, vb.net and teh work of mathematics this is about as far as I want to go (Try dividing 10 /3 in each and notice the slight difference)

I apologise again for some of this code. My only defence is that I have to cofde like this in order to fit in with the rest of the cross tabs.