Added an ability to define the columns in the find window grid nto be definable. This was more fiddly than I first thought as the find window column selection control was in fact a treeview - despite never using expand/collapse functionality. So I've had to replace the control and rework all the logic for add/insert/remove/move up/move down.
Because of the different control types I've removed the drag and drop functionality from the first tab - this may get noticed, or it may not - I'm not sure if anyone knows its there - I didn't. Just in case I've raised another task (TFS-12289)

Also, while I was there, I've resized the form to make it bigger as this was a bit of a bug-bear for a long time.