This now picks up on the licence setting as defined in System Manager.
Having the licence checks in a stored proc was always a bit of a security hole -as it was very easy for someone to edit the proc and just put  a return 1 at the bottom - bypassing the whole thing. Having the licence down in the code is tighter and cleaner.

The three added files have been cribbed from OpenHR web. Ideally we should probably look at a shared code folder to try and avoid this duplication, but we're a bit pressed for time at the moment.

There will be further check-ins for this task to remove/tidy up udfASRNetIsModuleLicensed and spASRIntActivateModule procs