Concurrency not taking into account users that have logged out.
The problem seems to be that two login processes can commence simulateously before the update to the licencehub can take place. A bit tricky to sort out completely becuase we need to make a valid database connection to work out what permissions we have and which of the DMI or SSI modules we need to navigate to first. All in all in means there's often a couple of seconds lag during which time it is possible for more than the licenced number of users to log in.

This changeset does not completely fix the issue but it does make it slightly better, by getting rid of an unecessary updateUsersOnlineCount callback. 

Also there were quirks whereby hitting an invalid licence when toggling between DMI and SSI was not counting down the correct user licences.

There will probably be subsequent changesets to betterize this even more!