This is a very simplistic framework for the absence triggers triggers.
Functionality that has been implemented in this framework is insert, update and delete of absences via Absence_Entry and Appointment_Absence_Entry. It splits the date range into individual rows (breakdown) and then merges it back into Absence. It handles subsequent updates of the entry rows and updates the corresponding Absence rows.

It does this by a mechanism of deletes and inserts (no updates) 

Functionality NOT implemented. 
1) Anything to do with the working pattern.
2) Inserting batch entries into the Entry tables.

It also currently relies on having some system generated triggers disabled as there is conflict in when the IDs are created, validation on the IDs (as we are creating child data first). It also requires the overnight job to be currently triggered.

This is obviously far from a finished product, but by running through the supplied test harness it should be possible to get a feel for how these triggers will perform. As a result this is not envisionaged to ever be viewed by QA in its current guise, but can be looked at by BA for a functionality overview.