Added some generic handling of errors if the hearthbeat or the sql service broker are non functional.
Creates an error model object and passes to ConfigurationErrors screen if the heartbeat and service broker are broken and bypasses the login form entirely.

I've created a styleBundle for openhrr and site.css because I couldn't get the linclude to work with vbhtml (maybe was just having an off day). Anyway bundles are cool.

Renamed the notificationhub to databasehub as its is generally used for database activity and can now be expanded to track the messaging from system manager *next check-ins)

The configuration error screen has a nice pretty broken picture on it, but I reckon someone with a better eye for colouring in could have a crack at making it a bit prettier (handle different resolutions, more message text perhaps)

Details of how to force a broken system are in the task description.