Absences were not correctly calculating the hours and day durations. Moved logic for day calculations to a udf for cleanliness of code. You will need to run the new script file for this to work.
There is a problem with having a combination of patterns in hours and day. It doubles up in teh merges absence table and, well, just gets a bit confused. I'm told in the real world this won't ever happen, but maybe we need to cater somehow?

Also it now fixes up the ID_1 rolling through into the merges absence column.