Misc bug fixes - all tied into one because of a significant overlap. Firstly the licence checks were a bit erratic because of the way we refresh the whole page when someone logs in. This means that SignalR is allocating a new connection ID during the login process, so we need to track the sessions and the connection id seperately.
Second part of this fix was that security manager could save changes even if users in that security group were logged in. I have re-instated the 5.2 logix which checks for the existence of users in the edited security group and blocks the save if the user is logged in. Because of SignalR we don't have to wait for the poll job to clear the session anymore.

On that point the signalR now takes over notification of the session timeout. So gone has pollMessage.aspx and its associated code. Session("timeout" and windows.timeout etc.

Users are now only notified of system saves  if they are logged in. (11801, 12570)

If a system administrator locks a database it no longer informs any logged in users, but instead removes the log in button and displays a message to the user.

Audit log entries are now made for all OpenHR web logins. If differenciates between OpenHR and Self-Service. (It doesn't show a log off if a user toggles between dmi/ssi after they have logged in, so in theory a user could log into dmi, switch to ssi and then logout and the log will show a login to dmi, a login to ssi and a log off ssi.

The concurrency check was a little erratic and would allow more than the licenced users. This is better than it was before, but its not mullet proof because of that damn annoying thing with a refresh everytime someone logs in. If your quick you can squeeze and extra user in if your timing is right. To fix this properly would mean a sledgehammer on the way we frame the login process, which would be a good thing and may be necessary as we cloud ready ourselves and performance becomes and issue.

Users logged into the web now appear as current logged in users in the System, Secuirty and Data Managers.

It goes with out saying that both updates scripts need running.