Changed text to OpenHR Customer Services.
The work item asks for all references to be changed, but this is quite tricky as some of the items are helfd in binary files (.frx) and there are also other numerous references to ASR Support, COA Support, Technical Support, Support, Helpdesk (you get the picture). 

Therefore this changeset merely addresses the licence key screen.