Absence calendar toggle options not working. Couple of problems where to do with different browsers. Firstly the way the region was processing meant some dates might not get evalulated correctly.
Secondly, the firing sequence of functions was not consistent between browsers, sorted with $(document).ready(function () {

And secondly, and this was bigger is that the bit that set the captions was reliant on innerText, which behaves differently and that the object beingh referenced did not always respond kindly to frmBlah.Date_xxx.something. Alter to reference this object via jQuery.