Removed the clear and ellipsis button for the template selection in mail merge.
This is far from ideal, however it wasn't working properly with Chrome or FireFox because these browsers don't allow us to view path information. So sledgehammer meet nut! We are likely to review hoe this functionality hangs together in future versions, but for the moment this is it.

Also, and I can't beleive QA haven't spotted this, added a colon after the Template label.