Size of string columns not honoured when running a custom report.
The beakaviour of the break could be up for discussion as it will split what appear to be identical values, but these would have been different is the specified column size was left alone (i.e. setup a column as size 4, and "Williams" and "Williamson" will all displa as "Will", but a page or break on change will still separate them. Not sure if this is correct, but its what DM does.

Note this changeset is part 1 of 2 (mail merge will be dealt with separately)