Preparation for consolidation (phase #4). Errors found when upgrading from 5.2
References to sp_ASRIntDefUsage created in the wrong order causing "The module will still be created; however, it cannot run successfully until the object exists." error  (spASRIntDefProperties) and 
Cannot drop the procedure 'dbo.spASRIntGetLinks', because it does not exist or you do not have permission.

Thus removed references to 
spASRIntDefProperties
spASRIntGetLinks

as these are already regenerated in the 8.1 script.