The working pattern array was problematic since its conversion from vb6 in v8.0 (namely zero-bounding being different between vb6 and vb.net). So I've taken the opportunity to simplify this section with a nice collection and using linq to grab the data we need.
Please note while reviewing this some comments in the technical description, that basically say when you start putting in absences that cross different working patterns you should not rely too heavily on the results from Data Manager as it would appear to have problems of its own.

This needs reviewing for database with static working pattersn and historic ones as the code flow is slightly different for both of these.