Changes as following code reviews and task item
Disable the forcibly disconnect if there are any fat clients logged in. Reworded as per task item

Display the difference between OpenHR Web and self-service users.

Audit the fact if they use the forcibly disconnect switch - this requires making the column larger in the database and in the grdAccess grid.

You will need to run the latest update script for these changes to work (overflow errors in the audit table otherwise)