There was some server side variables in the menu.js from when these functions used to sit in a .asp page. This was causing errors in oWasp scanning software.
Trying to remove them opened up a can of worms becuase the flow of the code was a bit rubbish (noticed task 8476 when implementing), so I've taken a bit of an axe to these functions and now implement as a ajax post rather than submitting the whole form and reloading the page. I can then get rid of a couple of action types (hardening for oWasp)

The messagebox has now been implemented as a modaldiv rather than an alert box.