The forgotten password text input was injectable. A couple of issues dealt with here, firstly I've removed the dependency of the .net system framework project (spadmin_commitresetpassword) as we can now use the heartbeat to achieve the same result - makes it a bit more secure that hoofing round usernames via stored procs - its all nice and sealed in a complied dll.
Second part of the fix is to implement a securestring for the password and use the sqlcredentials class for our connection string. The reason for this is that the connection string was failing if users started to put unusual charcaters in them (so now if a user decides to put 
';DELETE FROM [dbo].[ASRSysEventLog]--
as their password instead of it actually executing that code it will now set that as their password, and they will be able to log in with those credentials. (The fat clients will not, but we are not fixing those in this round)