Removing on error resume nexts and start of task af migrating on error goto to try catches.
In this changeset the absence calendar report has been migrated and the home controller actions no longer ignore errors. (This may cause funnynesses which we'll have to see what it does - and fix appropraitely rather than just IGNORING the error!

Moved the custom error pages so it is only shown in release mode.