Migrated the vb6 on errors to try catch. There was a large amount of redundant code in here which rather than spend time migrating over I just zapped it.
Added an extension code stub for strings to clean up any sensitive data that may be returned from the server. Will implement in a subsequent check-in.

I've kept the flow logic of the existing error handling, which isn't ideal, but I don't want to break anything at this late stage. There's a good case to break some of the huge functions into smaller ones and put a consistent try, catch throw sequence in.