Had to put the validateinput false on the data submit because we need to be able to cater for "dodgy" code when submitting recedit. I know this is a bit of a sledgehammer to crack a ball bearting, but its the only short term solution I can see.
I was hopeful of using the "exclude" parameter on validateinput, however this feature that was introduced in mvc3 was deprecated in mvc4 (how handy). A bit of Googling suggests two solutions 1: Write our own custom validation class, or simply put attributes on the individual properties on the model. (That does however mean converting it over to a mvc class model structure!)