This is a baby step approach to breaking up some of the monster data_submit and optiondata submit functions.
In this changeset I've moved out the bulk of the code from data_submit to seperate controller classes. Whilst none of the programatic flow logic had changed there has been a change to how the return data is sent back to the home controller - whereas before variables were altered directly it now passes resposnses through a response class. This avoids the need to use byref values in the function calls and makes it ready for when we want to call these controller actions directly.

Please bear in mind that this is a start of this task and there will be subsequent improvements. I loook forward to your comments. No seriously I do.

To test these code changes the areas that need looking at are saving records ij n recedit (both valid ahnd invalid attempts) and moving around the training booking functionality.