Task 15754: Created transforms for Web.config for Release and Debug modes
Test guidance:  Right click on Web.Release.config and from the popup select "Preview Transform"; the result should be the Web.config that will be shipped if the solutions is built in Release mode. do the same for Web.Debug.config. Make sure Web.config is the correct one depending on publishing and/or building the solution in Release or Debug mode