Removed frmOriginalDefinition from report running as this was unnecessary postable form that could be used for an xss attack.
txtDefn_Name and txtDefn_ErrMsg are used in util_run, but all the other inputs were unused (overhang from classic asp and client side activeX days)

I suspect the txtCancelPrint in cross tabs is no longer used, but I didn't want to start hacking out this section of code without a bit more time to play with.

DataOnlyPrint was not referenced anywhere in the code - so bye bye.