Started work on getting rid of optiondata submit controllers classes and forms. The basic logic is that we no longer rely on teh action type of a form, but instead create some json data and post directly into a controller classes whcih accepts as a modelled class.
The trick and the compliaction with these chjanges is that everything went through emption option and optiondata and used the same session variables for different actions. I've tried as part of this to tidy that up, but there's a risk some variables are now incorrectly populated which means certain things may not refresh. What's needed is a bit of a regression blast on the form that are affected. 

Areas affected that needs some sanity checking are: 
Training Booking
Quick Find
Lookups on rec edit
Recedit (LOad, Save, Copy, New)
Filter (and clear filter)
Links (training booking - course find, as an example)
Load/Save expression
Standard Reports, Load & Run

Added an anti forgery helper which is used when rendering this these post arrays in ascx and a bit of javascript that gets it from the main page when called from a .js file (because a client side bit of code has no way of generating one of the ids). Each of the models inherits from a base class which implements an anti-forgery property.

Started moving all the action properties from a combination fo strings and integers to a strongly typed OptionActionType, whcih is also duplicated as an enum down the .js side.