The following controller actions are only ever partial rendered and thus the controller action is never called. util_run_outputoptions util_run_customreports util_run_calendarreport_main util_run_customreportsMain
util_run: Is only ever called from from the prompted values controller, so merged in with that.

util_test_expression_submit: Was never called in the appliaction

WorkflowPendingSteps
the menu_autoLoadPage was only used for workflow pending steps, so renamed, rewritten to post a model to the controller class and then got rid of the validateinput. Had to add a no cache attribute so that data would refresh

util_run_workflow
Now passed by an ajax call and via a model, which is in turned wrapped into a session object and passed through to the view. Ideally I'd just pass the object directly into the view, however as util_run_workflow is an ascx page and not razor then it won't accept it. This is a bit of a hlfway solution that gives us a bit of readability without having to rewrite the form. (And it simplifies the amount of session variables in play)

util_run_promptedvalues - Just taken off, none of the parameters should ever have unsafe data in them.

There are some validateinputs remaining - test/validate expressions and recedit

Added autoencode to the defsel jqgrid so that its not injectable with data coming from the server.