WE have to recheck the permissions of the user on the server side becuase we cannot rely that they haven't fiddled on the client side and re-enabled buttons. This has meant a bit of harmonising on the code that builds the permissions because we sometimes relied on utility types, other times text. The code for the clienmt is now the same as the server. This does involve adding a utilitytype extension because the permission name is different from the associated enum (we need to get rid of the utlXXXX prefix -ts just a legacy from long ago.
Removed frmOnly mine and converted over to a ajax post data. This has mean we can clean up fromMenu and txtGotoOptionRecordID (which has always confused ame as to what the hell it did)

The defsels arte now called directly instead of through workarea refresh with a lovely little model.

A new error page has been added for invalid permissions - this should only ver be seen if a user "hacks" the button disabled status.