Security on buttons and accordian. The user can (if they are hacky-minded) enable buttons that they don't have security permissions on. We now evaluate the permissions again back on the server so that they can't get anything they shouldn't
This changeset was primarily designed to implement the security on the standard reports, but has scope crept to remove the ShowInReportFrame function because at the time I was having to call multiple controller classes. This has invloved tweaks to picklist. (At this point creating new and saving picklists seems to error, but this is not related - will fixup after this changeset)

I was getting numerous problems loading the standard reports, so I've puts some models into the controller classes. I haven't didied up and given the properties nice names, as this was too much for one changeset. I've also got rid of a few session variables and strongly typed some others.