Changed spASRIntGetRecord to read the table values directly into variables instead of via a temporary table.
Renamed the function while I was here as to try and get riud of the pesky (and slow) sp_xxxx naming convention.

Some background for your intelectual pleasure. The reason why we were pulling into a temp table and then into variables was that many moons ago SQL could only handle a maximum of 4000 characters in its nvarchar, and the executeSQL proc that we use to insert into variabnles had to use nvarchar. We got around this by creating a varchar (unlimited) to execute the statement and then simply read the resultant data. All this extra overhead is made redundant now that nvarchars can be defined as MAX.

While I was running the update script I had some errors with sp_ASRIntGetFindRecords, so I tidied those up at the same time.