Removed the last of the validateinput attributes. This now means we are using the default behaviour of asp.net and unless we've specially set and attribute to be marked as safe you should not be able to inject any unsafe javascript up to the controller actions.
This changeset deals with the epxression builder, validate, and testing, with and without prompted values.

Because of the way the submitform posts I've had to use a different tmpDialog as it needed to behave differently for training booking, and I thought it was safer just to branch off and have one exlusively for expressions. 

Switching to post means I can get rid of 3 forms - reducing our footprint just a little bit more.

Also got rid of the rather timesome and confusing util_dialog_expression - all it did was redierct to validate or test_prompted.

Work to do : Does not have the checks down on the component level, so you can add a value of <script>alert("hello ducky")</script> and you will get a potentially dangerous message. These controller actions need to be modelled. Next changeset maybe?

Also the allowHTML is not an ideal fix (its much better than validateinput!!!) as it is still open to some injectable code. The problem I've got at the moment is that javascript encodeURI and httputility.htmlencode and decode don't seem to be compatible. I think its something to do with the way the components1 string builds up from the treeview. Again, another changeset