Changed the expression component popup to use a model and hence we can now save/create expressions even with < ' & etc characters in the name and or contents of a component. Split away from the emptyoptions submit code and pass the information around directly using model classes (much cleaner). This has then meant I can start uncluttering the gotoOptionDataModel.
As the user can enter scriptable code (either through the web or the data manager) I've had to make some changes in the promptedvalues test and run to hncode the values properly.

Use of the AllowHtml is not ideal - we really need a clean way of encoding text down on the client before sending up to the controller, however the bit that gets values from the treeview is return something quite wacky and needs cleaning up. escape() and encodeUri doesn't work in tandem with httputility.encode and decode - which is a right royal pain in the...

To test, add/edit/insert some components - put some javascript/sql code, make sure expressions load, save, test and actually run when included in reports.