The usagechecks were getting stuck after deleting a utility.
Simplified the process od deleting by moving some of the code from the views to the controller class and utilise the definitionproperties model to populate the checkusages form.

I suppose I shopuld ahve done a bit of renaming and moved over to razer at the same time, but we have a pen test, and I'm in a hurry. Maybe something for later in the sprint? Please comment.