We've now added OLE columns to system defined orders. This was displayed system.byte[] error messages in grid that had not been coded to handle them (namely the find window). So we now specifically code the stored procs to ignore ole columns. Essentially this has affected all the training booking screens, anywhere where we have a lookup column defined (manager on Personnel_records for example), or have a link defined (delegate link on training booking for example)
We don't add a datatype filter everywhere as this would break the loops where security is evalutes (overcomplex merge on sys.syscolumns to get the permissions), so typically you will find the modified stored procs have two references to ASRSysOrderItems - the first with no filter, and the second one which builds the order collection has the filter.

Also of note is that I've added aliases to the modified part of the procs - 1 to increase readability, but mainly to help filter out the asrsysorderitems keyword when I had an awful lot in the project.

I've also had to modify the field component of expression builder because trying to add a binary column to an order could cause issues. You can get to this functionality by creating a report with a calculation refercing a child table with an order on it.

Also some unecessary stored procs were noted when looking for ASRSysOrderItems - removed these rather than "fixing" them.

Hopefully goes without saying that both update script will need running for these changes to work.