The working pending steps was exceptionally slow loading up (particularily func8.1 on qasrvr01). The reason for this was a pesky cursor in the pending steps stored proc. Tweaked it around a bit so that puts the data into a temp table and then uses the cursor to scroll through that. I looked at doing the whole lot in one select statement, but the spASRWorkflowStepDescription has all sorts of other procs hanging from it - and its all a bit complex.
This has required removing the sql7 checks. I'm not going to lose any sleep about someone asking to run workflow on sql7!

There are two associated tasks to this, although none quite match exactle the work here. 14953 is where the first performance degredation was noted (complete red herring). And then 16100 which turned off some caching. I'm pretty sure the performance drag has been there since day 1, but would only have been present in the login process, but the truning off the cache makes it very visible with a refresh and clicking on pending steps.

This should address the performance in all ways of it being called.