This is a bit of a patch rather than an elegant fix - the real problem is that when you add a new lookup table in system manager it does not populate the ASRSysViewMenuPermissions table, so I've just put an isnull in there.
When I say the real problem is in system manager, I'm actually lying - the REAL problem is that we store a whole table of permissions just for this setting. Good security practice dictates that by default everything is off unless you find some data to the contrary, but we've gone the other way - assume they have permissions unless there's a row to say the opposite. Ugly.

Anyway this is a quick fix to return the same results as Data Manager.