Changes so that only users affected by a security manager updated are notified and kicked out. This has re-instated the functionality that existed before 8.0 vhanges to the ado. Now the security manager will only notify users if something within their security group has changed (system or data permission - not necessarily addition or deletion of users).
To implement this I've added a module and a notification group the lock table which is populated by the security manager, but left blank by the system admin (you'll need to do a shift login for this to work). These functions are NOT included in the standard update script, but instead in the lock.sql which is called before any user checks are made. If you want a long winded reason for this come over and have a chat!)

There are then changes in the web. Upon login the client will connect to the databasehub with the current user group (groups are created automatically by signalR), and blank user groups don't cause any problems when called (that will happen when a user hits the login screen as its all embedded in the banner.ascx.

Moved the displaysignalr message to OpenHR for neatness.