You could add dodgy javascript code into the heading field of reports.
The reported problem (17634) in mail merge never displayed it, but kept it in the background with <calc>text in there which caused posting errors as we've got rid of the validateinput(false) flag (error 500). The fix for this element was to tidy up the stored proc.

To test try putting some dodgy code in th heading field (I like using <script>alert('hello ducky');</script> for an inmteresting effect.