Bug 16415: Setting a Record Filter breaks navigation away from Rec Edit or Find Window
After discussing with ME and PC we agreed that filters will now be kept per table.
Previously we had a single Session variable to keep the filter for the current table; this filter was removed everytime the find window was accessed either for a parent or child table.
Now a filter can be set for any number of parent tables and the filters will be kept until cleared; filters can also be kept for any number of child tables, but they will be cleared when the user navigates up to the find window of the parent table (or when cleared using the menu, as was always the case)

To Test
Set a filter for Personnel Records, and then edit any record and set a filter for any number of its child tables (Absence, Salary, etc.); navigating between the child tables keeps the filters; navigating to the parent Find window and its filter will still be set; editing any record and navigating to one of the child tables for which a filter was set previously, and the filter has been cleared