A user could un-disable buttons (email, purge, delete) in the event log and execute them without having valid permissions.
I've added and attribute which is now called before the controller action fires, this validates against the security permissions object. I was hoping to roll this change into defsel, however that has parameters which I've yet to figure out how to pass into the attribute class.

I've also renamed a coupld of the views as it was a tad confusing what forms were loaded and when and what they actually did.

Got rid of an unecessary form "frmEmail". It wasn't used anymore.

I've also moved new user functionality over to this mechanism