Runtime calculations are not stored away with the return type, therefore this has top be calculated when the expression is read. A couple of points to note here. Firstly a returntype and a column type are not one to one related - for example expressions have no concept of integers, thus we have to return as a numeric. Also table value components could return different data types. We also need to cater for byref calculations (e.g. Get field from database record).
This fix is purely for the loading/creating of report definitions, nothing to do with the runtime of the reports.

To test, we need to handle calculation of all return types making sure they convert to the nearest logical column type.
Prompted values convert to the nearest return type
Table Lookup values return the type of the return column.
Calculations within calculations recurse with the above logic.