Aplication vunerable to session fixation attacks. We now have to regenerate the seesion id once a user has logged in.
We have now split the login process into two chunks. The ValidateAntiForgeryToken has to be taken off the second login function because we will have reset the session id during this stage.

This has invloved tweaks to the signalr for the licence stuff as it maintains its own mapping between session ids and connection ids.