A redirecttoAction using a complex model renders the model information directly in the url. This was causing an issue displaying username and password in clear text if the user could not log in for whatever reason. So in order to avoid unecessarily passing round a view when redirecting I've now moved some of the checking from the actuallogin to the first login.
The one type of failed login this does not cover is if the user is a valid sql user, but does not have access to any of the web modules (A fat client user only is the only example I can think of). Fixing this would involve hacking apart the login stored procedure to handle the permissions and the logins differently - at this stage I'm trying to keep as simple as possible so that we can release 8.1

Just a bit of background. This login process is somewhat overcomplex because we need to clear the session when logging in as to avoid session hijacking (oWasp)