Problem : We have assigned a class 'absolutefull optiondatagridpage' to the div in which we are loading the grid content.    The same class used for the bulkbooking selection and for picklist selection.
This class is used to calculate the height & width of the grid. Now when we close the popup the height of the dialog becomes 0 and when opens the dialog again it does have the data but not showing because of the length having 0.

Solution : Added an Id to the picklist selection div and used the combination of id and class to calculate the height & widht of the grid.