Mail merge field names were being truncated.
As we are splitting/uploading the template file between the web and the fat client then we can now handle the two code streams differently. 
The DatMan will operate as before - automatically trim the merge points and append a 1, 2 etc. This still leaves it open to fun and games if you start putting more long calculations in - it'll be very difficult to tell what merge point relates to what column/calc. However this is the way its always been, and without completely changing the mail merge engine we use in vb6 then this is how it will have to stay. 
The web will now be fixed so that it does not automatically trim the mergefield names. These are the consequences of this decision.
1) Defining the template manually is a tinsy bit easier as you won't have to guess when to shorten/append a number to any of the field names.
2) Defining a fully working template with very long names could potentially break the DM if the same template was used for both, as the fat client is expecting shortened data - The users should be made aware that the two templates are different.
3) Creating a template through the data manager will give you a template with shortened names, this will need to be manually adjusted for use with the web version - Load template, right-click mergefield edit field and change the code.
4) Exisiting templates might not work straight out of the box, due to the shortened names - see point above.
After discussion with Mark E the above was considered acceptable as defining a template is typically a one off event and when its uploaded all run smoothly.